✨ NEW: Implements ColumnBuilder to build table by columns
