+++ src/_FullInternals.ts
@@ -14,36 +14,36 @@
 import { Pamela as Core } from './_CoreInternals';
 
 // shapes
-import { Arc } from './shapes/Arc';
-import { Arrow } from './shapes/Arrow';
-import { Circle } from './shapes/Circle';
-import { Ellipse } from './shapes/Ellipse';
-import { Image } from './shapes/Image';
-import { Label, Tag } from './shapes/Label';
-import { Line } from './shapes/Line';
-import { Path } from './shapes/Path';
-import { Rect } from './shapes/Rect';
-import { RegularPolygon } from './shapes/RegularPolygon';
-import { Ring } from './shapes/Ring';
-import { Sprite } from './shapes/Sprite';
+import { Arc }              from './shapes/Arc';
+import { Arrow }            from './shapes/Arrow';
+import { Circle }           from './shapes/Circle';
+import { Ellipse }          from './shapes/Ellipse';
+import { Image }            from './shapes/Image';
+import { Label, Tag }       from './shapes/Label';
+import { Line }             from './shapes/Line';
+import { Path }             from './shapes/Path';
+import { Rect }             from './shapes/Rect';
+import { RegularPolygon }   from './shapes/RegularPolygon';
+import { Ring }             from './shapes/Ring';
+import { Sprite }           from './shapes/Sprite';
 import { Star }             from './shapes/Star';
 import { GrowPolicy, Text } from './shapes/Text';
 import { TextPath }         from './shapes/TextPath';
-import { Transformer } from './shapes/Transformer';
-import { Wedge } from './shapes/Wedge';
+import { Transformer }      from './shapes/Transformer';
+import { Wedge }            from './shapes/Wedge';
 
 // filters
-import { Blur } from './filters/Blur';
-import { Brighten } from './filters/Brighten';
-import { Contrast } from './filters/Contrast';
-import { Emboss } from './filters/Emboss';
-import { Enhance } from './filters/Enhance';
-import { Grayscale } from './filters/Grayscale';
-import { HSL } from './filters/HSL';
-import { HSV } from './filters/HSV';
-import { Invert }       from './filters/Invert';
-import { Kaleidoscope } from './filters/Kaleidoscope';
-import { Mask }      from './filters/Mask';
+import { Blur }                     from './filters/Blur';
+import { Brighten }                 from './filters/Brighten';
+import { Contrast }                 from './filters/Contrast';
+import { Emboss }                   from './filters/Emboss';
+import { Enhance }                  from './filters/Enhance';
+import { Grayscale }                from './filters/Grayscale';
+import { HSL }                      from './filters/HSL';
+import { HSV }                      from './filters/HSV';
+import { Invert }                   from './filters/Invert';
+import { Kaleidoscope }             from './filters/Kaleidoscope';
+import { Mask }                     from './filters/Mask';
 import { Noise }                    from './filters/Noise';
 import { Pixelate }                 from './filters/Pixelate';
 import { Posterize }                from './filters/Posterize';
@@ -59,7 +59,7 @@ import { Point2D }                  from './common/Point2D';
 import { ExportVariable }           from './shapes/ExportVariable';
 import { CropTransformer }          from './shapes/CropTransformer';
 import { TableBuilder }             from './builders/TableBuilder';
-import { Row } from './builders/Row';
+import { RowBuilder }               from './builders/RowBuilder';
 
 export const Pamela = Core.Util._assign(Core, {
   Arc,
@@ -85,7 +85,7 @@ export const Pamela = Core.Util._assign(Core, {
   GrowPolicy,
   CropTransformer,
   TableBuilder,
-  Row,
+  RowBuilder,
   RichText,
   RichTextSource,
   Point2D,
