+++ src/index-types.d.ts
@@ -164,7 +164,7 @@ declare namespace Pamela {
   export type Table = import('./shapes/Table').Table;
   export type TableConfig = import('./shapes/Table').TableConfig;
   export type TableBuilder = import('./builders/TableBuilder').TableBuilder;
-  export type Row = import('./builders/Row').Row;
+  export type RowBuilder = import('./builders/RowBuilder').RowBuilder;
 
   export const Rect: typeof import('./shapes/Rect').Rect;
   export type Rect = import('./shapes/Rect').Rect;
